Fix for loop variables used with lengthof
lengthof returns type size_t, but most for loops used int as a loop
variable.  Fix that.  This avoids possible warnings about
signed/unsigned mismatches under higher warning levels.  (The compiler
will likely optimize these loops beyond recognition, so this shouldn't
affect the generated code much.)

Discussion: https://www.postgresql.org/message-id/flat/d639aede-209f-412b-927a-d38d4848b370%40eisentraut.org

+++ src/backend/main/main.c
@@ -243,7 +243,7 @@ main(int argc, char *argv[])
 DispatchOption
 parse_dispatch_option(const char *name)
 {
-	for (int i = 0; i < lengthof(DispatchOptionNames); i++)
+	for (size_t i = 0; i < lengthof(DispatchOptionNames); i++)
 	{
 		/*
 		 * Unlike the other dispatch options, "forkchild" takes an argument,
